Working collision detection?!

+++ include/constants.hpp
@@ -6,5 +6,5 @@ constexpr int WINDOW_HEIGHT = 1080 * 0.8;
 
 constexpr int PLAYER_SPEED = 400;
 constexpr int PLAYER_FRICTION = 5000;
-constexpr int MAX_PLAYER_SPEED = 350;
+constexpr float MAX_PLAYER_SPEED = 350.0f;
 } // namespace constants
